requirements.txt
commonmark==0.9.1
pkg_resources==0.0.0
Pygments==2.12.0
python-Levenshtein==0.12.2
PyYAML==6.0
rich==12.4.4
thefuzz==0.19.0
typing_extensions==4.1.1
